+++ Assets/GothicVR/Scenes/Bootstrap.unity
@@ -339,7 +339,7 @@ MonoBehaviour:
   ShowMusicLogs: 0
   EnableSounds: 1
   EnableMusic: 1
-  enableSoundCulling: 0
+  enableSoundCulling: 1
   vobCulling: 1
   vobCullingSmall:
     maxObjectSize: 1.2
@@ -954,6 +954,7 @@ GameObject:
   - component: {fileID: 1896514220}
   - component: {fileID: 1896514221}
   - component: {fileID: 1896514222}
+  - component: {fileID: 1896514227}
   - component: {fileID: 1896514225}
   - component: {fileID: 1896514223}
   - component: {fileID: 1896514226}
@@ -1134,6 +1135,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 9a2b6d5503112494f89b3a4c775c37d5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &1896514227
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1896514212}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d8a5433ca9664e58a7e90654db1d382c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1660057539 &9223372036854775807
