+++ Assets/GothicVR/Scripts/Manager/Culling/VobMeshCullingManager.cs
@@ -6,14 +6,14 @@ using GVR.Debugging;
 using GVR.Util;
 using UnityEngine;
 
-namespace GVR.Manager
+namespace GVR.Manager.Culling
 {
     /// <summary>
     /// CullingGroups are a way for objects inside a scene to be handled by frustum culling and occlusion culling.
     /// With this set up, VOBs are handled by camera's view and culling behaviour, so that the StateChanged() event disables/enables VOB GameObjects.
     /// @see https://docs.unity3d.com/Manual/CullingGroupAPI.html
     /// </summary>
-    public class CullingGroupManager : SingletonBehaviour<CullingGroupManager>
+    public class VobMeshCullingManager : SingletonBehaviour<VobMeshCullingManager>
     {
         // Stored for resetting after world switch
         private CullingGroup vobCullingGroupSmall;
